requirement.txt
python-multipart
fastapi
uvicorn
ultralytics
opencv-python-headless
pillow
torch
numpy